Attempting to add RECEIVE_WINDOW_OPENING_TIME to geom output data for using determining nadir position in surface return code.
Need to edit EDR and ANC parse functions. There seems to be an error in the way the ancilliary data is
initially parsed from the science data in read_EDR. Unphysical values are currently output for
RECEIVE_WINDOW_OPENINING_TIME. I believe that the read_Anc is functioning properly and that the error
lies in read_EDR alone.